# docs/api/rill/admin/v1/public.swagger.yaml
swagger: "2.0"
info:
  title: Rill API
  description: It is used to manage organizations, projects, deployments, users, and roles.
  version: v1
  contact:
    name: Rill Data
    url: https://rilldata.com
  license:
    name: Apache 2.0
    url: https://github.com/fridencao/stardata/blob/main/LICENSE.md
host: api.rilldata.com
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /v1/orgs:
    get:
      summary: ListOrganizations lists all the organizations currently managed by the admin
      operationId: AdminService_ListOrganizations
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1ListOrganizationsResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: pageSize
          in: query
          required: false
          type: integer
          format: int64
        - name: pageToken
          in: query
          required: false
          type: string
    post:
      summary: CreateOrganization creates a new organization
      operationId: AdminService_CreateOrganization
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1CreateOrganizationResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/v1CreateOrganizationRequest'
  /v1/orgs/{org}:
    get:
      summary: GetOrganization returns information about a specific organization
      operationId: AdminService_GetOrganization
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1GetOrganizationResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: org
          in: path
          required: true
          type: string
        - name: superuserForceAccess
          in: query
          required: false
          type: boolean
    delete:
      summary: DeleteOrganization deletes an organization
      operationId: AdminService_DeleteOrganization
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1DeleteOrganizationResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: org
          in: path
          required: true
          type: string
    patch:
      summary: UpdateOrganization deletes an organizations
      operationId: AdminService_UpdateOrganization
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1UpdateOrganizationResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: org
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              description:
                type: string
              newName:
                type: string
              displayName:
                type: string
              logoAssetId:
                type: string
              faviconAssetId:
                type: string
              thumbnailAssetId:
                type: string
              defaultProjectRole:
                type: string
              billingEmail:
                type: string
...
definitions:
...
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
...
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
...
  v1CreateOrganizationRequest:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      displayName:
        type: string
  v1CreateOrganizationResponse:
    type: object
    properties:
      organization:
        $ref: '#/definitions/v1Organization'
...
  v1DeleteOrganizationResponse:
    type: object
...
  v1GetOrganizationResponse:
    type: object
    properties:
      organization:
        $ref: '#/definitions/v1Organization'
      permissions:
        $ref: '#/definitions/v1OrganizationPermissions'
...
  v1ListOrganizationsResponse:
    type: object
    properties:
      organizations:
        type: array
        items:
          type: object
          $ref: '#/definitions/v1Organization'
      nextPageToken:
        type: string
...
  v1Organization:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
        title: Globally unique
      displayName:
        type: string
      description:
        type: string
      logoUrl:
        type: string
      faviconUrl:
        type: string
      thumbnailUrl:
        type: string
      customDomain:
        type: string
      defaultProjectRoleId:
        type: string
      quotas:
        $ref: '#/definitions/v1OrganizationQuotas'
      billingCustomerId:
        type: string
      paymentCustomerId:
        type: string
      billingEmail:
        type: string
      billingPlanName:
        type: string
      billingPlanDisplayName:
        type: string
      createdOn:
        type: string
        format: date-time
      updatedOn:
        type: string
        format: date-time
...
  v1OrganizationPermissions:
    type: object
    properties:
      admin:
        type: boolean
      guest:
        type: boolean
      readOrg:
        type: boolean
      manageOrg:
        type: boolean
      readProjects:
        type: boolean
      createProjects:
        type: boolean
      manageProjects:
        type: boolean
      readOrgMembers:
        type: boolean
      manageOrgMembers:
        type: boolean
      manageOrgAdmins:
        type: boolean
  v1OrganizationQuotas:
    type: object
    properties:
      projects:
        type: integer
        format: int32
      deployments:
        type: integer
        format: int32
      slotsTotal:
        type: integer
        format: int32
      slotsPerDeployment:
        type: integer
        format: int32
      outstandingInvites:
        type: integer
        format: int32
      storageLimitBytesPerDeployment:
        type: string
        format: int64
...
  v1UpdateOrganizationResponse:
    type: object
    properties:
      organization:
        $ref: '#/definitions/v1Organization'
...